Tree gathering and growing.

+++ Assets/Scripts/ScriptableObjectScripts/TreesSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Tree", menuName = "Tree")]
 public class TreesSO : ScriptableObject
 {
-    public float growTime;
+    public int growTime;
     public int woodAmount;
     public int cutCount;
 }
